Updated transaction history to render in a panel
