+++ app/components/TransactionHistory/TransactionHistoryPanel/Transactions.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@import '../../../styles/variables';
 
 .noTransactions {
     position: absolute;
